Added comments

+++ index.js
@@ -23,6 +23,7 @@ connection.connect((err) => {
   menuPrompt();
 });
 
+// Runs choice prompt on start and after each action
 function menuPrompt() {
   inquirer
     .prompt([
@@ -47,6 +48,7 @@ function menuPrompt() {
       },
     ])
     .then((response) => {
+      // runs the correct method based on user choice
       switch (response.type) {
         case "View employees":
           viewEmployees();
@@ -88,6 +90,7 @@ function menuPrompt() {
     });
 }
 
+// outputs the list of all employees along with the relevant information from their role and department
 function viewEmployees() {
   let query = `SELECT employee.id, employee.first_name, employee.last_name, title, department.name AS department, salary, CONCAT(manager.first_name, " ", manager.last_name) AS manager FROM
   employee INNER JOIN role ON employee.role_id = role.id 
@@ -103,20 +106,28 @@ function viewEmployees() {
   });
 }
 
+// Uses mysql to select all available roles and employees (to be a manager),
+// Then prompts the user to input a name for the new employee,
+// Then lets the user select their role and manager, if any
 function newEmployee() {
+  // Selects all available roles from the table
   connection.query("SELECT title, id FROM role", (err, res) => {
     if (err) throw err;
     let titles = res;
-
     //console.log(titles);
+    // Selects all available employees to be candidates for manager
+    // Could update table to have a value set for manager candidates
     connection.query(
       `SELECT CONCAT(first_name, " ", last_name) AS name, id FROM employee`,
       (err, res) => {
         if (err) throw err;
 
         let managers = res;
+        // Adds an option for no manager to be selected
         managers.unshift({ name: "None", id: 0 });
         //console.log(managers);
+
+        // Prompts the user for the information of the new employee
         inquirer
           .prompt([
             {
@@ -145,21 +156,26 @@ function newEmployee() {
           .then((response) => {
             //console.log(titles, managers);
 
+            // gets the id of the chosen role
             let roleId = titles.filter((el) => el.title === response.title)[0]
               .id;
 
+            // gets the id of the chosen role
             let managerId = managers.filter(
               (el) => el.name === response.manager
             )[0].id;
 
+            // makes an object containing the correct format for the table
             let emp = {
               first_name: response.first,
               last_name: response.last,
               role_id: roleId,
             };
+            // adds the manager id if one was selected
             if (managerId != 0) {
               emp.manager_id = managerId;
             }
+            // adds the employee to the table
             connection.query("INSERT INTO employee SET ?", emp, (err, res) => {
               if (err) throw err;
               console.log(
@@ -175,7 +191,9 @@ function newEmployee() {
   });
 }
 
+// Allows the user to delete an employee from the table
 function deleteEmployee() {
+  // Gets a list of all employees from the table
   connection.query(
     `SELECT CONCAT(first_name, " ", last_name) AS name, id FROM employee`,
     (err, res) => {
@@ -184,6 +202,8 @@ function deleteEmployee() {
       let employees = res;
       //managers.unshift({ name: "None", id: 0 });
       //console.log(managers);
+      // Presents the user with a list of employees to delete then
+      // Makes sure they want to delete them
       inquirer
         .prompt([
           {
@@ -202,32 +222,40 @@ function deleteEmployee() {
         ])
         .then((response) => {
           //console.log(titles, managers);
+          // Makes sure the user wants to delete someone
+          if (response.confirm) {
+            // Selects the id of the employee
+            let empId = employees.filter(
+              (el) => el.name === response.employee
+            )[0].id;
 
-          let empId = employees.filter((el) => el.name === response.employee)[0]
-            .id;
+            // Deletes the employee from the table
+            connection.query(
+              "DELETE FROM employee WHERE ?",
+              [{ id: empId }],
+              (err, res) => {
+                if (err) throw err;
+                console.log(`Removed ${response.employee} from the database`);
+                // Call updateProduct AFTER the INSERT completes
 
-          connection.query(
-            "DELETE FROM employee WHERE ?",
-            [{ id: empId }],
-            (err, res) => {
-              if (err) throw err;
-              console.log(`Removed ${response.employee} from the database`);
-              // Call updateProduct AFTER the INSERT completes
-
-              menuPrompt();
-            }
-          );
+                menuPrompt();
+              }
+            );
+          }
         });
     }
   );
 }
 
+// Updates the role of a selected user
 function updateRole() {
+  // Selects all the available roles in the table
   connection.query("SELECT title, id FROM role", (err, res) => {
     if (err) throw err;
     let titles = res;
 
     //console.log(titles);
+    // Selects all the available employees in the table
     connection.query(
       `SELECT CONCAT(first_name, " ", last_name) AS name, id FROM employee`,
       (err, res) => {
@@ -236,6 +264,7 @@ function updateRole() {
         let employees = res;
         //managers.unshift({ name: "None", id: 0 });
         //console.log(managers);
+        // Uses the mysql query responses to display options to the user
         inquirer
           .prompt([
             {
@@ -254,14 +283,16 @@ function updateRole() {
           ])
           .then((response) => {
             //console.log(titles, managers);
-
+            // Gets the id of the selected role
             let roleId = titles.filter((el) => el.title === response.title)[0]
               .id;
 
+            // gets the id of the selected employee
             let empId = employees.filter(
               (el) => el.name === response.employee
             )[0].id;
 
+            //sets the new role for the selected employee using the selected options
             connection.query(
               "UPDATE employee SET ? WHERE ?",
               [{ role_id: roleId }, { id: empId }],
@@ -281,6 +312,7 @@ function updateRole() {
   });
 }
 
+// Shows all the roles and their associated department
 function viewRoles() {
   let query = `SELECT 
   role.id, title, salary, name AS department
@@ -297,9 +329,13 @@ ORDER BY role.id`;
   });
 }
 
+// Lets the user select a department to add a role to,
+// Then lets the user enter a new role and salary
 function newRole() {
+  // Gets all the departmetns in the table
   connection.query("SELECT name, id FROM department", (err, res) => {
     if (err) throw err;
+    // Prompts the user to select a department then enter a new role and salary
     inquirer
       .prompt([
         {
@@ -322,14 +358,17 @@ function newRole() {
       .then((response) => {
         //console.log(titles, managers);
 
+        // gets the id of the department
         let deptId = res.filter((el) => el.name === response.department)[0].id;
 
+        // Makes an object to be added to the role table
         let dept = {
           title: response.role,
           salary: Number.parseFloat(response.salary),
           department_id: deptId,
         };
 
+        // adds the role to the table
         connection.query("INSERT INTO department SET ?", dept, (err, res) => {
           if (err) throw err;
           console.log(`Added ${dept.title} to the database`);
@@ -341,7 +380,10 @@ function newRole() {
   });
 }
 
+// Lets the user delete a role fro the table
+// Only roles with no employees assigned to them can be deleted
 function deleteRole() {
+  // Selects all the roles in the table as choices
   connection.query(`SELECT title, id FROM role`, (err, res) => {
     if (err) throw err;
 
@@ -367,37 +409,44 @@ function deleteRole() {
       .then((response) => {
         //console.log(titles, managers);
 
-        let roleId = roles.filter((el) => el.title === response.role)[0].id;
-        connection.query(
-          "SELECT COUNT(*) FROM employee GROUPBY role_id WHERE role_id = ?",
-          roleId,
-          (err, res) => {
-            if (err) throw err;
+        if (response.confirm) {
+          // Gets the id of the chosen role
+          let roleId = roles.filter((el) => el.title === response.role)[0].id;
+          // Checks if any employees have the chosen role
+          connection.query(
+            "SELECT COUNT(*) FROM employee GROUPBY role_id WHERE role_id = ?",
+            roleId,
+            (err, res) => {
+              if (err) throw err;
 
-            if (res.length > 0) {
-              console.log(
-                "You cannot delete a role while an employee is assigned to it"
-              );
-              menuPrompt();
-            } else {
-              connection.query(
-                "DELETE FROM role WHERE ?",
-                [{ id: roleId }],
-                (err, res) => {
-                  if (err) throw err;
-                  console.log(`Removed ${response.role} from the database`);
-                  // Call updateProduct AFTER the INSERT completes
+              // If an employee has the role it tells the user and sends them to the menu
+              if (res.length > 0) {
+                console.log(
+                  "You cannot delete a role while an employee is assigned to it"
+                );
+                menuPrompt();
+                // Otherwise it removes the role from the table
+              } else {
+                connection.query(
+                  "DELETE FROM role WHERE ?",
+                  [{ id: roleId }],
+                  (err, res) => {
+                    if (err) throw err;
+                    console.log(`Removed ${response.role} from the database`);
+                    // Call updateProduct AFTER the INSERT completes
 
-                  menuPrompt();
-                }
-              );
+                    menuPrompt();
+                  }
+                );
+              }
             }
-          }
-        );
+          );
+        }
       });
   });
 }
 
+// Selects all the departments to show the user
 function viewDepartments() {
   let query = `SELECT * FROM department ORDER BY id`;
   connection.query(query, (err, res) => {
@@ -409,6 +458,7 @@ function viewDepartments() {
   });
 }
 
+// Lets the user create new departments
 function newDepartment() {
   inquirer
     .prompt([
@@ -435,7 +485,9 @@ function newDepartment() {
     });
 }
 
+// Allows the user to delete a department if no user works in it
 function deleteDepartment() {
+  // Selects all the departments to show the users
   connection.query(`SELECT name, id FROM department`, (err, res) => {
     if (err) throw err;
 
@@ -460,48 +512,53 @@ function deleteDepartment() {
       ])
       .then((response) => {
         //console.log(titles, managers);
-
-        let deptId = depts.filter((el) => el.name === response.dept)[0].id;
-        connection.query(
-          `SELECT COUNT(*) FROM
+        if (response.confirm) {
+          // Gets the id of the selected department
+          let deptId = depts.filter((el) => el.name === response.dept)[0].id;
+          // Checks to make sure no employees work in the deartment
+          connection.query(
+            `SELECT COUNT(*) FROM
           employee INNER JOIN role ON employee.role_id = role.id 
           INNER JOIN department ON role.department_id = department.id 
           LEFT JOIN employee AS manager ON employee.manager_id = manager.id
           WHERE department_id = ?
          GROUP BY department.id;`,
-          deptId,
-          (err, res) => {
-            if (err) throw err;
+            deptId,
+            (err, res) => {
+              if (err) throw err;
 
-            if (res.length > 0) {
-              console.log(
-                "You cannot delete a department while an employee is assigned to it"
-              );
-              menuPrompt();
-            } else {
-              connection.query(
-                "DELETE FROM department WHERE ?",
-                [{ id: deptId }],
-                (err, res) => {
-                  if (err) throw err;
-                  console.log(`Removed ${response.dept} from the database`);
-                  // Call updateProduct AFTER the INSERT completes
+              // If any employees work in that department, tell the user and send them to the menu
+              if (res.length > 0) {
+                console.log(
+                  "You cannot delete a department while an employee is assigned to it"
+                );
+                menuPrompt();
+                // Otherwise delete the department from the table
+              } else {
+                connection.query(
+                  "DELETE FROM department WHERE ?",
+                  [{ id: deptId }],
+                  (err, res) => {
+                    if (err) throw err;
+                    console.log(`Removed ${response.dept} from the database`);
+                    // Call updateProduct AFTER the INSERT completes
 
-                  menuPrompt();
-                }
-              );
+                    menuPrompt();
+                  }
+                );
+              }
             }
-          }
-        );
+          );
+        }
       });
   });
 }
 
+// Gets the budget of each created department where the budget is the sum of all employee salaries
 function departmentBudgets() {
   let query = `SELECT department.name, SUM(salary) AS budget FROM
   employee INNER JOIN role ON employee.role_id = role.id 
-  INNER JOIN department ON role.department_id = department.id 
-  LEFT JOIN employee AS manager ON employee.manager_id = manager.id
+  RIGHT JOIN department ON role.department_id = department.id 
  GROUP BY department.name
  ORDER BY SUM(salary) DESC`;
